added X11/GLX support (Linux)
