+++ gfx.h
@@ -3,7 +3,7 @@
 
 #include "entity.h"
 
-void gfx_init(void *);
+void gfx_init();
 void gfx_kill(void);
 void gfx_clear(void);
 void gfx_render_entity(struct entity *);
